requirements.txt
atomicwrites==1.3.0
attrs==19.1.0
Click==7.0
Flask==1.1.1
importlib-metadata==0.18
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
more-itertools==7.2.0
numpy==1.16.4
packaging==19.0
pandas==0.25.0
pluggy==0.12.0
py==1.8.0
pyparsing==2.4.1.1
pytest==5.0.1
python-dateutil==2.8.0
pytz==2019.1
six==1.12.0
unicode==2.7
wcwidth==0.1.7
Werkzeug==0.15.5
zipp==0.5.2
